py_scripts/requirements.txt
adblock==0.6.0
asciidoc==10.2.1
colorama==0.4.6
Jinja2==3.1.4
MarkupSafe==3.0.2
Pygments==2.18.0
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6-WebEngine==6.8.0
PyQt6-WebEngine-Qt6==6.8.1
PyQt6_sip==13.9.1
python-rofi==1.0.1
PyYAML==6.0.2
-e git+https://github.com/qutebrowser/qutebrowser.git@1bcb536a44351ea5d2a0eec6d51a247a97555118#egg=qutebrowser
setuptools==75.6.0
wheel==0.45.1
adblock==0.6.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asciidoc==10.2.1
attrs==24.3.0
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
distro==1.9.0
docstring-parser==0.15
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
instructor==0.6.8
Jinja2==3.1.4
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
openai==1.58.1
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6-WebEngine==6.8.0
PyQt6-WebEngine-Qt6==6.8.1
PyQt6_sip==13.9.1
python-rofi==1.0.1
PyYAML==6.0.2
-e git+https://github.com/qutebrowser/qutebrowser.git@1bcb536a44351ea5d2a0eec6d51a247a97555118#egg=qutebrowser
rich==13.9.4
setuptools==75.6.0
shell_gpt==1.4.4
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.12.2
wheel==0.45.1
yarl==1.18.3
